Move from any-sketch fingerprinter to common-cpp
Move from any-sketch/src/main/cc/any_sketch/fingerprinters to
common-cpp/src/main/cc/common-cpp/fingerprinters
