+++ src/main/cc/wfanet/panelmatch/common/crypto/BUILD.bazel
@@ -83,9 +83,9 @@ cc_library(
     ],
     strip_include_prefix = _INCLUDE_PREFIX,
     deps = [
-        "@any_sketch//src/main/cc/any_sketch/fingerprinters",
         "@com_github_glog_glog//:glog",
         "@com_google_absl//absl/status",
         "@tink_base//cc/util:secret_data",
+        "@wfa_common_cpp//src/main/cc/common_cpp/fingerprinters",
     ],
 )
